# Conflicts:
#	src/main/java/poomasi/global/error/BusinessError.java

+++ src/main/java/poomasi/domain/image/service/ImageService.java
@@ -7,6 +7,12 @@ import poomasi.domain.image.dto.ImageRequest;
 import poomasi.domain.image.entity.Image;
 import poomasi.domain.image.entity.ImageType;
 import poomasi.domain.image.repository.ImageRepository;
+import poomasi.domain.image.validation.ImageOwnerValidator;
+import poomasi.domain.image.validation.ImageOwnerValidatorFactory;
+import poomasi.domain.member._profile.entity.MemberProfile;
+import poomasi.domain.member._profile.service.MemberProfileService;
+import poomasi.domain.member.entity.Member;
+import poomasi.domain.member.repository.MemberRepository;
 import poomasi.global.error.BusinessException;
 
 import java.util.Date;
@@ -20,18 +26,49 @@ import static poomasi.global.error.BusinessError.*;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ImageService {
+
+    private static final int DEFAULT_IMAGE_LIMIT = 5;
+    private static final int MEMBER_PROFILE_IMAGE_LIMIT = 1;
+
     private final ImageRepository imageRepository;
+    private final ImageOwnerValidatorFactory validatorFactory;
+    private final MemberRepository memberRepository;
+    private final MemberProfileService memberProfileService;
+
 
     @Transactional
-    public Image saveImage(ImageRequest imageRequest) {
+    public Image saveImage(Long memberId, ImageRequest imageRequest) {
         // 기존 이미지가 있는 경우 복구 또는 예외 처리 (실제 복구 로직과는 차이가 있음)
+        validateImageOwner(memberId, imageRequest.type(), imageRequest.referenceId());
         validateImageLimit(imageRequest);
 
         Image image = findExistingOrRecoverableImage(imageRequest)
                 .map(existingImage -> recoverImageOrThrow(existingImage, imageRequest))
                 .orElseGet(() -> imageRequest.toEntity(imageRequest));
 
+        if (imageRequest.type() == ImageType.MEMBER_PROFILE) {
+            linkImageToMemberProfile(imageRequest.referenceId(), image);
+        }
+
         return imageRepository.save(image);
+    }
+
+    // 이미지 주인이 맞는지 검증
+    private void validateImageOwner(Long memberId, ImageType type, Long referenceId) {
+        // 관리자는 제외
+        if (isAdmin(memberId)) {
+            return;
+        }
+        ImageOwnerValidator validator = validatorFactory.getValidator(type);
+        if (validator != null && !validator.validateOwner(memberId, referenceId)) {
+            throw new BusinessException(IMAGE_OWNER_MISMATCH);
+        }
+    }
+
+    private boolean isAdmin(Long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new BusinessException(MEMBER_NOT_FOUND));
+        return member.isAdmin();
     }
 
     private Optional<Image> findExistingOrRecoverableImage(ImageRequest imageRequest) {
@@ -50,21 +87,34 @@ public class ImageService {
     }
 
     private void validateImageLimit(ImageRequest imageRequest) {
-        if (imageRepository.countByTypeAndReferenceIdAndDeletedAtIsNull(imageRequest.type(), imageRequest.referenceId()) >= 5) {
+        int imageLimit = DEFAULT_IMAGE_LIMIT;
+        if (imageRequest.type() == ImageType.MEMBER_PROFILE) {
+            imageLimit = MEMBER_PROFILE_IMAGE_LIMIT; // 멤버 프로필 이미지는 한 장으로 제한
+        }
+
+        if (imageRepository.countByTypeAndReferenceIdAndDeletedAtIsNull(imageRequest.type(), imageRequest.referenceId()) >= imageLimit) {
             throw new BusinessException(IMAGE_LIMIT_EXCEED);
         }
     }
 
+    private void linkImageToMemberProfile(Long referenceId, Image savedImage) {
+        MemberProfile memberProfile = memberProfileService.getMemberProfileById(referenceId);
+        memberProfile.setProfileImage(savedImage);
+        memberProfileService.saveMemberProfile(memberProfile);
+    }
+
     // 여러 이미지 저장
     @Transactional
-    public List<Image> saveMultipleImages(List<ImageRequest> imageRequests) {
+    public List<Image> saveMultipleImages(Long memberId, List<ImageRequest> imageRequests) {
         return imageRequests.stream()
-                .map(this::saveImage)
+                .map(imageRequest -> saveImage(memberId, imageRequest))
                 .collect(Collectors.toList());
     }
 
     @Transactional
-    public void deleteImage(Long id) {
+    public void deleteImage(Long memberId, Long id) {
+        Image image = getImageById(id);
+        validateImageOwner(memberId, image.getType(), image.getReferenceId());
         imageRepository.deleteById(id);
     }
 
@@ -79,8 +129,9 @@ public class ImageService {
 
     // 이미지 수정
     @Transactional
-    public Image updateImage(Long id, ImageRequest imageRequest) {
+    public Image updateImage(Long memberId, Long id, ImageRequest imageRequest) {
         Image image = getImageById(id);
+        validateImageOwner(memberId, image.getType(), image.getReferenceId());
 
         if (!image.getType().equals(imageRequest.type()) ||
                 !image.getReferenceId().equals(imageRequest.referenceId())) {
@@ -93,9 +144,9 @@ public class ImageService {
     }
 
     @Transactional
-    public void recoverImage(Long id) {
-        Image image = imageRepository.findById(id)
-                .orElseThrow(() -> new BusinessException(IMAGE_NOT_FOUND));
+    public void recoverImage(Long memberId, Long id) {
+        Image image = getImageById(id);
+        validateImageOwner(memberId, image.getType(), image.getReferenceId());
 
         if (image.getDeletedAt() == null) {
             throw new BusinessException(IMAGE_ALREADY_EXISTS);
@@ -106,4 +157,5 @@ public class ImageService {
         image.setDeletedAt(null);
         imageRepository.save(image);
     }
+
 }
